Re-thinking auto-insert-as.
Now there is acommand that adds as close to the next to be scripted
command. Finally found a way to make this close to correct for
destruct and induction by using "!" modifier on the destructed names,
so that the automatic naming does not reuse the this name.

Probably still very approximated for inversion.