Fail silently if Coq's version can't be detected
Rationale: if Coq isn't installed, we will detect it when trying to run
it, and we don't need to duplicate the error logic.

Additionally, change from using process-lines to using shell-command,
possibly through file-name-handler. The reason for this change is that
we want to do the version detection on the remote server if we're
running in Tramp.