Fixing #147 and #91 + others indentation bugs.
While fixing these I discovered several flaws in indentation (what a
suprise). The probem is the following: since we don't have a precise
grammar of tactics, smie often decides that the parent of a sub-part
of a tactic is the previous command instead of the current tacic.

Typical example (fixed now but in general):

             foo.
  apply bar
             with bar'.

Since "apply ... bar'" is considered as one leaf of the grammar, it is
considered to be a child of the previous dot.

     .
     /\
    /  \
  foo  apply...bar'

Therefore indentation of "with" wants to align with parent "." and
hence with "foo".

Basically we should try to define a much more precise grammar of
complex tactics (with with, as, using etc) in order to fix the
problem. Of course this has the drawback of making impossible to
support user tactic notations.